[0.3.9] Refactoring of parser architecture:
- Basic parser moved to the new [base] folder;
- Dummy implementation of AbstractParser, the parent class to all parsers, introduced;
- Polymorphism unit test extended with a method that returns another method, made sure this is polymorphic as I am going to need that specifically in my parser inheritance implementation.
