+++ scripts/parsers/base/abstract_parser.py
@@ -1,0 +1,4 @@
+import argparse
+
+class AbstractParser:
+  pass
