+++ scripts/parsers/base/basic_throwing_parser.py
