+++ test/exploratory/oop/test_polymorphism.py
@@ -7,6 +7,12 @@ class ParentClass:
   def overloaded_method(self):
     return "CALL: parent"
 
+  def overloaded_returns(self):
+    return self._returns_impl()
+
+  def _returns_impl(self):
+    return "returns parent..."
+
 class ChildClass(ParentClass):
   def __init__(self):
     super(ChildClass, self).__init__()
@@ -15,6 +21,9 @@ class ChildClass(ParentClass):
 
   def overloaded_method(self):
     return "CALL: child"
+
+  def _returns_impl(self):
+    return "returns child..."
 
 class TestPolymorphism(unittest.TestCase):
   def setUp(self):
@@ -27,6 +36,10 @@ class TestPolymorphism(unittest.TestCase):
     self.assertEqual("parent", self.the_child.parents_backed_up_param)
     self.assertEqual("child", self.the_child.param)
     self.assertEqual("CALL: child", self.the_child.overloaded_method())
+
+  def test_return_is_polymorphic(self):
+    self.assertEqual("returns parent...", self.the_parent.overloaded_returns())
+    self.assertEqual("returns child...", self.the_child.overloaded_returns())
 
   def tearDown(self):
     pass
